Day4-6/Day12/class11/pyproject.toml
[tool.poetry]
name = "class11"
version = "0.1.0"
description = ""
authors = ["AMZeeshan Haider <shani.pcr2244@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
psycopg = {extras = ["binary"], version = "^3.2.4"}
sqlmodel = "^0.0.22"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
